+++ API/Models/Box.cs
@@ -5,7 +5,7 @@ namespace API.Models
         public int Id { get; set; }
         public required string GId { get; set; }        
         public required string SerialNumber { get; set; } 
-        public int InBoxRemaining { get; set; }
+        public double InBoxRemaining { get; set; }
         public DateTime ExpiryDate { get; set; }
         
         public ICollection<DispensingLog> DispensingLogs { get; set; } = new List<DispensingLog>();
